tests kinda sorta work
