+++ src/libraries/Microsoft.Bcl.Cryptography/tests/Microsoft.Bcl.Cryptography.Tests.csproj
@@ -26,14 +26,14 @@
              Link="CommonTest\System\Security\Cryptography\MLKemAlgorithmTests.cs" />
     <Compile Include="$(CommonTestPath)System\Security\Cryptography\MLKemTestData.cs"
              Link="CommonTest\System\Security\Cryptography\MLKemTestData.cs" />
+    <Compile Include="$(CommonTestPath)System\Security\Cryptography\MLKemImplementationTests.cs"
+             Link="CommonTest\System\Security\Cryptography\MLKemImplementationTests.cs" />
     <Compile Include="$(CommonTestPath)System\Security\Cryptography\MLKemTests.cs"
              Link="CommonTest\System\Security\Cryptography\MLKemTests.cs" />
-    <Compile Include="$(CommonTestPath)System\Security\Cryptography\MLKemTests.ArgValidation.cs"
-             Link="CommonTest\System\Security\Cryptography\MLKemTests.ArgValidation.cs" />
-    <Compile Include="$(CommonTestPath)System\Security\Cryptography\MLKemTests.Keys.cs"
-             Link="CommonTest\System\Security\Cryptography\MLKemTests.Keys.cs" />
-    <Compile Include="$(CommonTestPath)System\Security\Cryptography\MLKemTests.Encapsulation.cs"
-             Link="CommonTest\System\Security\Cryptography\MLKemTests.Encapsulation.cs" />
+    <Compile Include="$(CommonTestPath)System\Security\Cryptography\MLKemBaseTests.cs"
+             Link="CommonTest\System\Security\Cryptography\MLKemBaseTests.cs" />
+    <Compile Include="$(CommonTestPath)System\Security\Cryptography\MLKemKeyTests.cs"
+             Link="CommonTest\System\Security\Cryptography\MLKemKeyTests.cs" />
     <Compile Include="$(CommonTestPath)System\Security\Cryptography\SP800108HmacCounterKdfTests.ArgValidation.cs"
              Link="CommonTest\System\Security\Cryptography\SP800108HmacCounterKdfTests.ArgValidation.cs" />
     <Compile Include="$(CommonTestPath)System\Security\Cryptography\SP800108HmacCounterKdfTests.Functional.cs"
