+++ src/libraries/System.Security.Cryptography/tests/MLKemOpenSslTests.Unsupported.cs
@@ -1,0 +1,16 @@
+// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+
+using Xunit;
+
+namespace System.Security.Cryptography.Tests
+{
+    public sealed class MLKemOpenSslTests
+    {
+        [Fact]
+        public static void MLKemOpenSsl_NotSupportedOnNonUnixPlatforms()
+        {
+            Assert.Throw<PlatformNotSupportedException>(() => new MLKemOpenSsl(new SafeEvpPKeyHandle()));
+        }
+    }
+}
