+++ src/libraries/System.Security.Cryptography/tests/System.Security.Cryptography.Tests.csproj
@@ -154,16 +154,16 @@
              Link="ProductionCode\Common\System\Net\StreamBuffer.cs" />
     <Compile Include="$(CommonTestPath)System\Security\Cryptography\MLKemAlgorithmTests.cs"
              Link="CommonTest\System\Security\Cryptography\MLKemAlgorithmTests.cs" />
+    <Compile Include="$(CommonTestPath)System\Security\Cryptography\MLKemBaseTests.cs"
+             Link="CommonTest\System\Security\Cryptography\MLKemBaseTests.cs" />
     <Compile Include="$(CommonTestPath)System\Security\Cryptography\MLKemTestData.cs"
              Link="CommonTest\System\Security\Cryptography\MLKemTestData.cs" />
+    <Compile Include="$(CommonTestPath)System\Security\Cryptography\MLKemImplementationTests.cs"
+             Link="CommonTest\System\Security\Cryptography\MLKemImplementationTests.cs" />
     <Compile Include="$(CommonTestPath)System\Security\Cryptography\MLKemTests.cs"
              Link="CommonTest\System\Security\Cryptography\MLKemTests.cs" />
-    <Compile Include="$(CommonTestPath)System\Security\Cryptography\MLKemTests.ArgValidation.cs"
-             Link="CommonTest\System\Security\Cryptography\MLKemTests.ArgValidation.cs" />
-    <Compile Include="$(CommonTestPath)System\Security\Cryptography\MLKemTests.Keys.cs"
-             Link="CommonTest\System\Security\Cryptography\MLKemTests.Keys.cs" />
-    <Compile Include="$(CommonTestPath)System\Security\Cryptography\MLKemTests.Encapsulation.cs"
-             Link="CommonTest\System\Security\Cryptography\MLKemTests.Encapsulation.cs" />
+    <Compile Include="$(CommonTestPath)System\Security\Cryptography\MLKemKeyTests.cs"
+             Link="CommonTest\System\Security\Cryptography\MLKemKeyTests.cs" />
     <Compile Include="$(CommonTestPath)System\Security\Cryptography\SP800108HmacCounterKdfTests.ArgValidation.cs"
              Link="CommonTest\System\Security\Cryptography\SP800108HmacCounterKdfTests.ArgValidation.cs" />
     <Compile Include="$(CommonTestPath)System\Security\Cryptography\SP800108HmacCounterKdfTests.Functional.cs"
@@ -416,7 +416,6 @@
     <Compile Include="KmacXof256Tests.cs" />
     <Compile Include="Length32Hash.cs" />
     <Compile Include="MD5Tests.cs" />
-    <Compile Include="MLKemOpenSslTests.cs" />
     <Compile Include="OidTests.cs" />
     <Compile Include="OidCollectionTests.cs" />
     <Compile Include="OpenSslNamedKeysTests.manual.cs" />
@@ -610,12 +609,13 @@
              Link="Common\Interop\Unix\System.Security.Cryptography.Native\Interop.ERR.cs" />
     <Compile Include="$(CommonPath)Interop\Unix\System.Security.Cryptography.Native\Interop.EVP.Kem.cs"
              Link="Common\Interop\Unix\System.Security.Cryptography.Native\Interop.EVP.Kem.cs" />
-    <Compile Include="$(CommonPath)Interop\Unix\System.Security.Cryptography.Native\Interop.EvpPkey.cs"
-             Link="Common\Interop\Unix\System.Security.Cryptography.Native\Interop.EvpPkey.cs" />
+    <Compile Include="$(CommonPath)Interop\Unix\System.Security.Cryptography.Native\Interop.EvpPkey.FromData.cs"
+             Link="Common\Interop\Unix\System.Security.Cryptography.Native\Interop.EvpPkey.FromData.cs" />
     <Compile Include="$(CoreLibSharedDir)System\IO\PersistedFiles.Unix.cs"
              Link="Common\System\IO\PersistedFiles.Unix.cs" />
     <Compile Include="$(CoreLibSharedDir)System\IO\PersistedFiles.Names.Unix.cs"
              Link="Common\System\IO\PersistedFiles.Names.Unix.cs" />
+    <Compile Include="MLKemOpenSslTests.Unix.cs" />
     <Compile Include="X509Certificates\HostnameMatchTests.Unix.cs" />
     <Compile Include="X509Certificates\X509FilesystemTests.Unix.cs" />
     <Compile Include="X509Certificates\X509StoreTests.Unix.cs" />
